+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [Feature] Remove the validation of the number of files in `BatchWriteDetectOperator`
 - [Feature] Rename operator `BatchWriteProcessOperator` to `BatchWriteProcessNdjsonOperator`
 - [Feature] Create new operator `BatchWriteProcessOrcOperator` that write ORC file format
+- [Feature] Create new operator `BatchWriteProcessedMoveOperator`
 
 **v0.0.52**
 - [Feature] Allow `BatchWriteDetectOperator` to receive a prefix to process only a few files